docker-compose/slackbot/requirements.txt
sqlalchemy
psycopg2-binary
slackclient
requests